+++ tests/conftest.py
@@ -6,39 +6,86 @@ import dotenv
 from unittest.mock import Mock
 
 from src.api.tagging.dto_mapping import is_live
+from src.fetch.model import FetcherConfig
 from src.tags.tagstore.abstract import Tagstore
 from src.tags.tagstore.filesystem_tagstore import FilesystemTagStore
 from src.tags.tagstore.rest_tagstore import RestTagstore
 from src.tagging.fabric_tagging.model import FabricTaggerConfig
-from src.tags.conversion import TagConverter, TagConverterConfig
 from src.common.content import Content, ContentConfig, ContentFactory
 from src.tagging.uploading.config import ModelUploadArgs, UploaderConfig, TrackArgs
 
 dotenv.load_dotenv()
 
-@pytest.fixture
-def test_qid():
-    return "iq__3C58dDYxsn5KKSWGYrfYr44ykJRm"
+"""Fabric Integration Test Fixtures"""
 
 @pytest.fixture
-def test_qid2():
-    return "iq__cebzuQ8BqsWZyoUdnTXCe23fUgz"
-
-@pytest.fixture
-def live_q():
-    token = os.getenv("LIVE_AUTH")
-    if not token:
-        pytest.skip("LIVE_AUTH not set in environment")
-    cfg = ContentConfig(
+def content_config() -> ContentConfig:
+    return ContentConfig(
         config_url="https://host-76-74-29-5.contentfabric.io/config?self&qspace=main", 
         parts_url="http://192.168.96.203/config?self&qspace=main",
         live_media_url="https://host-76-74-29-5.contentfabric.io/config?self&qspace=main"
     )
-    factory = ContentFactory(cfg=cfg)
-    q = factory.create_content(qhit="iq__467CAS4BvPQ39go6aLmX6v3ZaTwD", auth=token)
+
+@pytest.fixture
+def qfactory(content_config) -> ContentFactory:
+    factory = ContentFactory(cfg=content_config)
+    return factory
+
+@pytest.fixture
+def qid():
+    return "iq__3C58dDYxsn5KKSWGYrfYr44ykJRm"
+
+@pytest.fixture
+def q(qfactory, qid):
+    auth_token = os.getenv("TEST_AUTH")
+    if not auth_token:
+        pytest.skip("TEST_AUTH not set in environment")
+
+    return qfactory.create_content(qhit=qid, auth=auth_token)
+
+@pytest.fixture
+def qid_legacy():
+    return "iq__cebzuQ8BqsWZyoUdnTXCe23fUgz"
+
+@pytest.fixture
+def q_legacy(qfactory, qid_legacy):
+    auth_token = os.getenv("TEST_AUTH")
+    if not auth_token:
+        pytest.skip("TEST_AUTH not set in environment")
+
+    return qfactory.create_content(qhit=qid_legacy, auth=auth_token)
+
+@pytest.fixture
+def qid_live():
+    return "iq__467CAS4BvPQ39go6aLmX6v3ZaTwD"
+
+@pytest.fixture
+def q_live(qfactory, qid_live):
+    token = os.getenv("LIVE_AUTH")
+    if not token:
+        pytest.skip("LIVE_AUTH not set in environment")
+
+    q = qfactory.create_content(qhit=qid_live, auth=token)
+
     if not is_live(q):
         pytest.skip("livestream is not running")
+
     return q
+
+@pytest.fixture
+def qid_assets():
+    return "iq__4BT8BBNEEDvysXqjZgj4BRA5jVo2"
+
+@pytest.fixture
+def q_assets(qfactory, qid_assets):
+    auth_token = os.getenv("ASSETS_AUTH")
+    if not auth_token:
+        pytest.skip("ASSETS_AUTH not set in environment")
+
+    return qfactory.create_content(qhit=qid_assets, auth=auth_token)
+
+
+"""Basic utility fixtures"""
 
 @pytest.fixture
 def temp_dir():
@@ -47,107 +94,12 @@ def temp_dir():
     yield temp_path
     shutil.rmtree(temp_path, ignore_errors=True)
 
-@pytest.fixture
-def media_dir(temp_dir: str) -> str:
-    """Create a media directory for testing"""
-    media_path = os.path.join(temp_dir, "media")
-    os.makedirs(media_path, exist_ok=True)
-    return media_path
+@pytest.fixture()
+def static_dir():
+    test_dir = os.path.join(os.path.dirname(os.path.abspath(__file__)), 'test-stuff')
+    return test_dir
 
-@pytest.fixture
-def tagger_config(media_dir) -> FabricTaggerConfig:
-    return FabricTaggerConfig(
-        media_dir=media_dir, 
-        uploader=UploaderConfig(
-            model_params={
-                "caption": ModelUploadArgs(
-                    default=TrackArgs(name="object_detection", label="Object Detection")
-                ),
-                "asr": ModelUploadArgs(
-                    default=TrackArgs(name="speech_to_text", label="Speech to Text"),
-                    overrides={
-                        "pretty": TrackArgs(name="auto_captions", label="Auto Captions")
-                    }
-                ),
-            }
-        ),
-    )
-
-@pytest.fixture
-def tag_converter() -> TagConverter:
-    """Create a TagConverter with test configuration"""
-    config = TagConverterConfig(
-        interval=5,  # 5-minute buckets for testing
-        name_mapping={
-            "object": "Object Detection",
-            "asr": "Speech to Text",
-            "shot": "Shot Detection"
-        },
-        single_tag_tracks=[],
-        coalesce_tracks=["asr"],
-        max_sentence_words=10
-    )
-    return TagConverter(config)
-
-@pytest.fixture
-def qfactory() -> ContentFactory:
-    cfg = ContentConfig(
-        config_url="https://host-154-14-185-98.contentfabric.io/config?self&qspace=main", 
-        parts_url="http://192.168.96.203/config?self&qspace=main",
-        live_media_url="https://host-76-74-34-204.contentfabric.io/config?self&qspace=main"
-    )
-    factory = ContentFactory(cfg=cfg)
-    return factory
-
-@pytest.fixture
-def writable_q(qfactory):
-    """Check if TEST_AUTH and TEST_QWT are set in environment"""
-    auth_token = os.getenv("TEST_AUTH")
-    write_token = os.getenv("TEST_QWT")
-    
-    if not write_token or not auth_token:
-        yield None
-        return
-
-    q = qfactory.create_content(qhit=write_token, auth=auth_token)
-
-    yield q
-    q.replace_metadata(metadata_subtree="video_tags", metadata={})
-    
-@pytest.fixture
-def q(qfactory, test_qid):
-    """Check if TEST_AUTH is set in environment"""
-    auth_token = os.getenv("TEST_AUTH")
-    
-    if not auth_token:
-        return Mock(qid=test_qid, qhit=test_qid)
-
-    return qfactory.create_content(qhit=test_qid, auth=auth_token)
-
-@pytest.fixture
-def q2(qfactory, test_qid2):
-    """Check if TEST_AUTH is set in environment"""
-    auth_token = os.getenv("TEST_AUTH")
-    
-    if not auth_token:
-        return Mock(qid=test_qid2, qhit=test_qid2)
-
-    return qfactory.create_content(qhit=test_qid2, auth=auth_token)
-
-"""
-@pytest.fixture
-def q(writable_q, readonly_q, test_qid):
-    #Create Content object with write token from environment
-    auth_token = os.getenv("TEST_AUTH")
-    write_token = os.getenv("TEST_QWT")
-
-    if write_token:
-        return writable_q
-    elif auth_token:
-        return readonly_q
-    else:
-        return Mock(qid=test_qid, qhit=test_qid)
-"""
+"""Tagstore Fixtures"""
 
 @pytest.fixture
 def rest_tagstore(q: Content) -> RestTagstore:
@@ -176,3 +128,14 @@ def tag_store(rest_tagstore: RestTagstore, filesystem_tagstore: FilesystemTagSto
         return rest_tagstore
     else:
         return filesystem_tagstore
+    
+"""Fetcher Fixtures"""
+
+@pytest.fixture
+def fetcher_config() -> FetcherConfig:
+    """Create a FetcherConfig for testing"""
+    
+    return FetcherConfig(
+        author="tagger",
+        max_downloads=4
+    )
