+++ tests/core_tagging/test_resource_manager.py
