+++ tests/integration_private_keys.json
@@ -1,0 +1,5 @@
+{
+    "mgm": "",
+    "eluvio": "",
+    "redbull": ""
+}
