+++ tests/test-processor/build.sh
